+++ tests/datasets/plugins/openeye/spice/mock_data/script.py
@@ -6,7 +6,7 @@ if __name__ == "__main__":
     dl_manager = datasets.DownloadManager()
 
     h5_file_path = dl_manager.download_and_extract(
-        "https://github.com/openmm/spice-dataset/releases/download/1.1.1/SPICE.hdf5",
+        "https://zenodo.org/records/8222043/files/SPICE-1.1.4.hdf5?download=1",
     )
 
     h5_file = h5py.File(h5_file_path, "r")
@@ -38,5 +38,32 @@ if __name__ == "__main__":
             "formation_energy",
             data=h5_file[key]["formation_energy"][:2],
         )
-
+        new_h5_file[key].create_dataset(
+            "mbis_dipoles",
+            data=h5_file[key]["mbis_dipoles"][:2],
+        )
+        new_h5_file[key].create_dataset(
+            "mbis_quadrupoles",
+            data=h5_file[key]["mbis_quadrupoles"][:2],
+        )
+        new_h5_file[key].create_dataset(
+            "mbis_octupoles",
+            data=h5_file[key]["mbis_octupoles"][:2],
+        )
+        new_h5_file[key].create_dataset(
+            "scf_dipole",
+            data=h5_file[key]["scf_dipole"][:2],
+        )
+        new_h5_file[key].create_dataset(
+            "scf_quadrupole",
+            data=h5_file[key]["scf_quadrupole"][:2],
+        )
+        new_h5_file[key].create_dataset(
+            "mayer_indices",
+            data=h5_file[key]["mayer_indices"][:2],
+        )
+        new_h5_file[key].create_dataset(
+            "wiberg_lowdin_indices",
+            data=h5_file[key]["wiberg_lowdin_indices"][:2],
+        )
     new_h5_file.close()